\name{connectToNetworkManager}
\alias{connectToNetworkManager}
\title{connectToNetworkManager}
\usage{
  connectToNetworkManager(port = port, timeout = timeout,
    maxWait = 10)
}
\arguments{
  \item{port}{}

  \item{timeout}{}

  \item{maxWait}{}
}
\description{
  connectToNetworkManager
}
\author{
  Barnet Wagman
}
\keyword{internal}

