\name{conToPF}
\alias{conToPF}
\title{Wrapper for socketConnection(), so we can catch exceptions.}
\usage{
  conToPF(port, timeout)
}
\arguments{
  \item{port}{}

  \item{timeout}{}
}
\description{
  Wrapper for socketConnection(), so we can catch
  exceptions.
}
\author{
  Barnet Wagman
}
\keyword{internal}

