\name{getClassPath}
\alias{getClassPath}
\title{Uses getJavaAppClassPath to get the class paths for cloudRmpi and cloudRmpiJars}
\usage{
  getClassPath()
}
\value{
  class path containing cloudRmpi jars and third party
  jars.
}
\description{
  Uses getJavaAppClassPath to get the class paths for
  cloudRmpi and cloudRmpiJars
}
\author{
  Barnet Wagman
}
\keyword{internal}

