\name{getClassPath.v0}
\alias{getClassPath.v0}
\title{getClassPath.v0
Uses getJavaAppClassPath to get the class paths for cloudRmpi and cloudRmpiJars}
\usage{
  getClassPath.v0()
}
\value{
  class path containing cloudRmpi jars and third party
  jars.
}
\description{
  getClassPath.v0 Uses getJavaAppClassPath to get the class
  paths for cloudRmpi and cloudRmpiJars
}
\author{
  Barnet Wagman
}
\keyword{internal}

