\name{getConPortForwardingApp}
\alias{getConPortForwardingApp}
\title{Gets existing connection or connects if possible.
If a new connection
is made, it is assign to conPortForwardingApp in .GlobalEnv as
well as being returned.}
\usage{
  getConPortForwardingApp(port, timeout, maxWait = 10)
}
\arguments{
  \item{port}{}

  \item{timeout}{}

  \item{maxWait}{}
}
\description{
  Gets existing connection or connects if possible. If a
  new connection is made, it is assign to
  conPortForwardingApp in .GlobalEnv as well as being
  returned.
}
\author{
  Barnet Wagman
}
\keyword{internal}

