\name{launchPortForwardingApp}
\alias{launchPortForwardingApp}
\title{launchPortForwardingApp}
\usage{
  launchPortForwardingApp(port, timeout, verbose)
}
\arguments{
  \item{port}{}

  \item{timeout}{}

  \item{verbose}{}
}
\description{
  launchPortForwardingApp
}
\author{
  Barnet Wagman
}
\keyword{internal}

