\name{ppe.getRunningNetworkNames}
\alias{ppe.getRunningNetworkNames}
\title{Get running network names}
\usage{
  ppe.getRunningNetworkNames()
}
\value{
  character vector of network names.
}
\description{
  Get the names of ppe-ompi networks owned by the user that
  have status='running'.
}
\author{
  Barnet Wagman
}

