\name{ppe.shutdownPortForwardingApp}
\alias{ppe.shutdownPortForwardingApp}
\title{Shut down the port forwarding Java app.}
\usage{
  ppe.shutdownPortForwardingApp(maxWait = 2)
}
\arguments{
  \item{maxWait}{maxiumum time (in seconds) to wait for a
  connection to the port forwarding application.}
}
\description{
  This function is called by .Last, to it is not usually
  necessary to use it directly.
}
\author{
  Barnet Wagman
}

