\name{sendPPEManagerCmd}
\alias{sendPPEManagerCmd}
\title{launchRReServerAccessApp}
\usage{
  sendPPEManagerCmd(con, cmd, maxWaitSecs = -1)
}
\arguments{
  \item{con}{}

  \item{cmd}{}

  \item{maxWaitSecs}{}
}
\value{
  reply message
}
\description{
  launchRReServerAccessApp
}
\author{
  Barnet Wagman
}
\keyword{internal}

